// types.ts
import { FunctionComponent } from "react";
import { IconType } from "react-icons";
export interface Service {
  id: string;
  Icon: IconType;
  title: string;
  about: string;
}

export interface Skill {
  id: string;
  Icon: IconType;
  name: string;
  level: string;
}

export interface IProject {
  id: number;
  name: string;
  description: string;
  image_path: string;
  deployed_url: string;
  github_url: string;
  category: Category[];
  key_techs: string[];
}

export type Category =
  | "react"
  | "node"
  | "next"
  | "django"
  | "mongo"
  | "firebase"
  | "commercejs"
  | "stripe";